refactor(config-ui): move components/text-tooltip to components/tooltip/text-tooltip
