+++ config-ui/src/components/tooltip/index.ts
@@ -16,15 +16,4 @@
  *
  */
 
-export * from './loading';
-export * from './divider';
-export * from './page-header';
-export * from './selector';
-export * from './dialog';
-export * from './table';
-export * from './toast';
-export * from './logo';
-export * from './card';
-export * from './inspector';
-export * from './action';
-export * from './tooltip';
+export * from './text-tooltip';
